src/backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Credit Insight backend"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "dotenv-python>=0.0.1",
    "opengauss-sqlalchemy>=2.2.0",
]
